Add files via upload
Linear regression is a popular statistical technique used to model and analyze the relationship between a dependent variable and one or more independent variables. In Python, linear regression can be implemented using libraries such as pandas, numpy, and scikit-learn.

In essence, linear regression aims to find the best-fitting line that represents the linear relationship between the independent variables (also known as features or predictors) and the dependent variable (also known as the target variable or response variable). The line is determined by estimating the coefficients or weights associated with each independent variable. These coefficients indicate the impact or contribution of each variable on the target variable.